+++ src/Converter.php
@@ -4,6 +4,10 @@ declare(strict_types=1);
 
 namespace Money;
 
+use InvalidArgumentException;
+
+use function sprintf;
+
 /**
  * Provides a way to convert Money to Money in another Currency using an exchange rate.
  */
@@ -21,11 +25,42 @@ final class Converter
 
     public function convert(Money $money, Currency $counterCurrency, int $roundingMode = Money::ROUND_HALF_UP): Money
     {
-        $baseCurrency = $money->getCurrency();
-        $ratio        = $this->exchange->quote($baseCurrency, $counterCurrency)->getConversionRatio();
+        return $this->convertAgainstCurrencyPair(
+            $money,
+            $this->exchange->quote(
+                $money->getCurrency(),
+                $counterCurrency
+            ),
+            $roundingMode
+        );
+    }
 
-        $baseCurrencySubunit    = $this->currencies->subunitFor($baseCurrency);
-        $counterCurrencySubunit = $this->currencies->subunitFor($counterCurrency);
+    /** @return array{0: Money, 1: CurrencyPair} */
+    public function convertAndReturnWithCurrencyPair(Money $money, Currency $counterCurrency, int $roundingMode = Money::ROUND_HALF_UP): array
+    {
+        $pair = $this->exchange->quote(
+            $money->getCurrency(),
+            $counterCurrency
+        );
+
+        return [$this->convertAgainstCurrencyPair($money, $pair, $roundingMode), $pair];
+    }
+
+    public function convertAgainstCurrencyPair(Money $money, CurrencyPair $currencyPair, int $roundingMode = Money::ROUND_HALF_UP): Money
+    {
+        if (! $money->getCurrency()->equals($currencyPair->getBaseCurrency())) {
+            throw new InvalidArgumentException(
+                sprintf(
+                    'Expecting to convert against base currency %s, but got %s instead',
+                    $money->getCurrency()->getCode(),
+                    $currencyPair->getBaseCurrency()->getCode()
+                )
+            );
+        }
+
+        $ratio                  = $currencyPair->getConversionRatio();
+        $baseCurrencySubunit    = $this->currencies->subunitFor($currencyPair->getBaseCurrency());
+        $counterCurrencySubunit = $this->currencies->subunitFor($currencyPair->getCounterCurrency());
         $subunitDifference      = $baseCurrencySubunit - $counterCurrencySubunit;
 
         $ratio = Number::fromString($ratio)
@@ -34,6 +69,6 @@ final class Converter
 
         $counterValue = $money->multiply($ratio, $roundingMode);
 
-        return new Money($counterValue->getAmount(), $counterCurrency);
+        return new Money($counterValue->getAmount(), $currencyPair->getCounterCurrency());
     }
 }